www/requirements.txt
bottle==0.12.8
numpy==1.9.2
pyparsing==2.0.3
python-dateutil==2.4.1
pytz==2014.10
six==1.9.0
matplotlib==1.4.3